Fixed a bunch of 64-bit warnings on Windows, and a few genuine potential bugs on other 64-bit platforms.

+++ RuleTableToTree.cpp
@@ -116,7 +116,7 @@ void shrink() {
    nodeseq = 0 ;
    curndd = recreate(oseq, curndd, neighborhood_size) ;
    cerr << "Shrunk from " << oseq.size() << " to " << seq.size() << endl ;
-   shrinksize = seq.size() * 2 ;
+   shrinksize = (int)(seq.size() * 2) ;
 }
 void write_ndd() {
    shrink() ;
